requirements.txt
opencv-python-headless
numpy
pyside6
onnxruntime
insightface
pillow
pyyaml
pyinstaller
pytest
